refactor: move screen components from App.tsx to src/ui/screens/
