+++ src/ui/screens/index.ts
@@ -1,2 +1,5 @@
 export { MainMenu } from './MainMenu';
 export { NewGame } from './NewGame';
+export { GameScreen } from './GameScreen';
+export { GameOverScreen } from './GameOverScreen';
+export { PlaceholderScreen } from './PlaceholderScreen';
